moving anything that uses log into the @fab/cli package makes a lot more sense, actually
