+++ packages/cli/src/errors/BuildFailed.ts
