+++ packages/cli/src/errors/InvalidConfig.ts
