+++ packages/cli/src/errors/InvalidPlugin.ts
