+++ packages/cli/src/errors/index.ts
@@ -1,0 +1,4 @@
+export * from './BuildFailed'
+export * from './InvalidConfig'
+export * from './InvalidPlugin'
+export * from './MissingConfig'
