+++ packages/cli/src/helpers/index.ts
@@ -1,0 +1,7 @@
+import chalk from 'chalk'
+
+export const log = {
+  error(str: string) {
+    console.log(chalk.red(str))
+  },
+}
