+++ packages/cli/src/index.ts
@@ -1,1 +1,3 @@
-export {run} from '@oclif/command'
+export { run } from '@oclif/command'
+
+export * from './errors'
