--- packages/core/runtime.js
@@ -1,1 +1,0 @@
-export * from './esm/runtime'
