--- packages/core/src/errors/index.ts
@@ -1,4 +1,0 @@
-export * from './MissingConfig'
-export * from './InvalidConfig'
-export * from './BuildFailed'
-export * from './InvalidPlugin'
