+++ packages/core/src/index.ts
@@ -1,7 +1,14 @@
+/*
+ * @fab/core: Type definitions and helper methods
+ *
+ * Be careful what you import in this project, everything needs to be Rollup-compatible
+ * and runnable inside a FAB, since FAB runtime components will import this.
+ *
+ * For build-time-specific code, use @fab/cli
+ *
+ * */
+
 export * from './types'
 export * from './constants'
 export * from './helpers'
-
-export * from './errors'
-
 export * from './models/ProtoFab'
