+++ packages/rewire-assets/src/build.ts
@@ -5,13 +5,10 @@ import {
   RewireAssetsArgs,
   RewireAssetsMetadata,
 } from './types'
-import {
-  filenameOutsideFabLocations,
-  InvalidConfigError,
-  getContentType,
-} from '@fab/core'
+import { filenameOutsideFabLocations, getContentType } from '@fab/core'
 import hasha from 'hasha'
 import path from 'path'
+import { InvalidConfigError } from '@fab/cli'
 
 export async function build(
   args: RewireAssetsArgs,
